requirements.txt
certifi==2020.6.20
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
json5==0.9.1
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.19.0
PyJWT==1.7.1
PySocks==1.7.1
python-dotenv==0.14.0
pytz==2019.3
requests==2.24.0
sendgrid==6.4.6
six==1.15.0
torchvision==0.3.0
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-linux_x86_64.whl
twilio==6.44.0
urllib3==1.25.9
Werkzeug==1.0.1